scalar

+++ scalar/src/lib.rs
@@ -7,10 +7,10 @@ pub fn diff(a: i32, b: i32) -> i32 {
 }
 
 pub fn pro(a: i8, b: i8) -> i8 {
- a*b
+    a*b
 }
 
-pub fn quo(a: i32, b: i32) -> i32 {
+pub fn quo(a: f32, b: f32) -> f32 {
     a/b
 }
